requirements.txt
asgiref==3.5.2
backoff==2.2.1
Deprecated==1.2.13
Django==3.2.16
googleapis-common-protos==1.57.0
grpcio==1.50.0
grpcio-tools==1.48.2
opentelemetry-api==1.14.0
opentelemetry-exporter-otlp-proto-grpc==1.14.0
opentelemetry-instrumentation==0.35b0
opentelemetry-instrumentation-django==0.35b0
opentelemetry-instrumentation-grpc==0.35b0
opentelemetry-instrumentation-wsgi==0.35b0
opentelemetry-proto==1.14.0
opentelemetry-sdk==1.14.0
opentelemetry-semantic-conventions==0.35b0
opentelemetry-util-http==0.35b0
protobuf==3.20.3
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.4.0
wrapt==1.14.1
